WIP
so it looks like the checkrb refresh function shouldn't have been used in some places, in rom c here this was breaking last folder of rom opening. so we need to back and see where it was added elsewhere and see if it was appropriate. we also have individual mask support for mm, that should be applied for pumkin tower and mm revival hacks. now with mk3 added it should be clear to end user what version of GLideN64 to use for the newer zelda hacks. so let's deal with checkrb refresh usage, then get the cheat support implemented and then that will be fc 5.3