ok this should fix rom list showing when its not supposed to
the function checkrb refresh is supposed to: show rom list, if on browser then refresh, and then if not browser then go back to hiding browser to help fix end emulation when browser off. but show rom list sometimes should not be called in some instances, like if you set last folder of rom opening for your rom directory as then it would show rom list as it selects the directory blocking the game view. now in these cases of recent rom directory and rom directory, i think if browser was off then it would be unecessarily doing show rom list and then going back into non browser mode, which is COMPLETELY a waste, so this should fix that