referencing 1.6 plus, maybe we can get dma a little better
https://github.com/pj64team/Project64-1.6-Plus/commit/31c94a6cc29b072328211027f181b2a956fcea73

accoording to this commit, this fixes "ai shogi garbles screens". dezaemon and stadium should already be fixed on our side so those changes are illrevent, and the taz express change is known to make some other games break. so lets test these changes (and what is that default for exactly?) and see how this impacts ai shogi 3 tomorrow when i wake up.