BETA v4
I removed delay RSP before because I thought it was unused, turns out it is very important for an extremely small percentage of games to boot properly (only NFL quarterback club that I know of so far). the previous code I added for checking savesate was not good, the idea was that oot hacks which borrow internal name for jabo per game hacks can run into a situation where you attempt to load state into another hack with the borrowed name, but the issue with this is that it interferes with plugin hot swap (you check for 4mb because oot hacks use 4mb, only the debug one's need 8!). several other fixes made as well from this commit and the previous commits between this one and beta v3.