Working on the Finale
Update to star road with lakitu cam fixed, various db improvements, make cheats fully english for consistency (cheats themselvs are english only), dim and don't save rom options when interpreter CPU is selected to be in use which aren't affected by said options, added more documentation in user guide, minor improvements.

I still want to setup counter factor 1 support for as many OoT hacks as possible (but need to be careful about this, as counter factor 1 is known to crash in OoT due to HLE video plugin pause fixes). I think I should also re-sort the language hader and numbering of the language files, just for cosmetic/be numbered correctly. this is a WIP and I may continue to think of more things to be changed for the "finale" release.,