WIP 09 now ready - almost perfect to me!
As of WIP 09, the emu is almost perfect, with the only thing I really feel that needs to be done next is audio stuff. Once that's taken care of, and once I look at whatever bugs get found while people use WIP 09, I think it will be perfect to me. 🙂