more mm cheat stuff
I decided to restore the magic line, since the hearts start filled the magic should too for consistency and technically it's not completely a 100% file if it's not ALL THE WAY maxed. I ported some more cheats over from the U region to J 1.0/1.1. That covers most of them except bombers hideout password, rapid fire, rock sirloin, invincibility, and ISG. title screen map select too in the case of J 1.1.